Move state to Booklist
